Add BetAnalyzerService. Add Dtos for MakeBetCommand. Add initial version of MakeBet functionality. Test MakeBet functionality on provided strings.

+++ VirtualRoulette.Domain/Entities/Jackpot.cs
@@ -18,7 +18,7 @@ namespace VirtualRoulette.Domain.Entities
             UpdatedAt = DateTime.UtcNow;
         }
 
-        public void AddToJackpot(decimal amount)
+        public void AddToJackpot(long amount)
         {
             Amount = Amount.Add(amount);
             UpdatedAt = DateTime.UtcNow;
